backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-chess>=1.999
aiosqlite>=0.19.0
pydantic>=2.0.0
